+++ CHANGELOG.md
@@ -7,7 +7,8 @@ unreleased
 
 ### Improvements
 
-- [beta] Add help page about the set replacement.
+- [beta] Rename “Set Replacement” feature to “Multiple Replacement.” 
+- [beta] Add help page about the multiple replacement.